requirements.txt
absl-py
tensorflow
opencv-python==4.1.0.25
https://dl.google.com/coral/python/tflite_runtime-2.1.0.post1-cp37-cp37m-macosx_10_14_x86_64.whl
numpy
Pillow
Cython
Matplotlib